requirements.txt
# Web Framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.0
Flask-Talisman>=1.1.0
Werkzeug>=3.0.0

# Machine Learning & Data
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0
scipy>=1.12.0

# Utilities
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0